requirements.txt
# RubyBet Production Dependencies
aiogram>=3.0.0
aiohttp>=3.9.0
python-dotenv
PyJWT>=2.8.0
bcrypt>=4.0.0
asyncpg>=0.29.0
PyNaCl>=1.5.0
solana>=0.30.0
